+++ opal/middleware.py
@@ -1,13 +1,14 @@
 """
 Opal Middlewares
 """
+from django.utils.deprecation import MiddlewareMixin
 # Hat tip:
 # http://kevinzhang.org/posts/django-angularjs-and-csrf-xsrf-protection.html
 
 ANGULAR_HEADER_NAME = 'HTTP_X_XSRF_TOKEN'
 
 
-class AngularCSRFRename(object):
+class AngularCSRFRename(MiddlewareMixin):
     def process_request(self, request):
         if ANGULAR_HEADER_NAME in request.META:
             token = request.META[ANGULAR_HEADER_NAME]
@@ -15,6 +16,6 @@ class AngularCSRFRename(object):
             del request.META[ANGULAR_HEADER_NAME]
 
 
-class DjangoReversionWorkaround(object):
+class DjangoReversionWorkaround(MiddlewareMixin):
     def process_request(self, request):
-        access = request.user.is_authenticated()  # noqa:
+        access = request.user.is_authenticated  # noqa: